Title: 산타의 선물 공장, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 250223/산타의 선물 공장/santa-gift-factory.cpp
@@ -90,7 +90,7 @@ int main() {
             int w_max;
             cin >> w_max;
 
-            int ret = 0;
+            long long ret = 0;
             int idx = 0;
             while (parent[idx] != idx) idx = beltnext[idx];
